wip: remove unused function for now, update rowmult with formula used in paper

+++ modules/qec/qec.h
@@ -48,7 +48,7 @@ public:
 
 	uint64_t get_x_stab(uint32_t i, uint32_t j);
 	uint64_t get_z_stab(uint32_t i, uint32_t j);
-	uint_fast8_t get_phase(uint32_t i);
+	uint_fast16_t get_phase(uint32_t i);
 
 	PackedInt64Array x_stabs();
 	PackedInt64Array z_stabs();
